# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Review App API",
    "version": "1.0.0",
    "description": "API documentation for the Software Design Artifacts Review Application",
    "contact": {
      "name": "API Support",
      "email": "support@reviewapp.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/reviews": {
      "post": {
        "summary": "Submit or update a review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "artifactType",
                  "artifactId",
                  "rating"
                ],
                "properties": {
                  "artifactType": {
                    "type": "string",
                    "enum": [
                      "requirement",
                      "story",
                      "activityDiagram",
                      "useCaseDiagram",
                      "sequenceDiagram",
                      "classDiagram",
                      "designPattern",
                      "mockup"
                    ]
                  },
                  "artifactId": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted successfully"
          },
          "400": {
            "description": "Invalid artifact type"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to review this artifact"
          },
          "404": {
            "description": "Artifact not found"
          }
        }
      }
    },
    "/api/reviews/my/{artifactType}/{artifactId}": {
      "get": {
        "summary": "Get current user's review for an artifact",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "artifactType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "requirement",
                "story",
                "activityDiagram",
                "useCaseDiagram",
                "sequenceDiagram",
                "classDiagram",
                "designPattern",
                "mockup"
              ]
            },
            "description": "Type of artifact"
          },
          {
            "in": "path",
            "name": "artifactId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the artifact"
          }
        ],
        "responses": {
          "200": {
            "description": "User's review for the artifact"
          },
          "400": {
            "description": "Invalid artifact type"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/reviews/{artifactType}/{artifactId}": {
      "get": {
        "summary": "Get all reviews for an artifact",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "artifactType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "requirement",
                "story",
                "activityDiagram",
                "useCaseDiagram",
                "sequenceDiagram",
                "classDiagram",
                "designPattern",
                "mockup"
              ]
            },
            "description": "Type of artifact"
          },
          {
            "in": "path",
            "name": "artifactId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the artifact"
          }
        ],
        "responses": {
          "200": {
            "description": "All reviews for the artifact"
          },
          "400": {
            "description": "Invalid artifact type"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to view reviews for this artifact"
          },
          "404": {
            "description": "Artifact not found"
          }
        }
      }
    },
    "/api/reviews/general-comment": {
      "post": {
        "summary": "Save general comment for artifact type",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "projectId",
                  "artifactType",
                  "comment"
                ],
                "properties": {
                  "projectId": {
                    "type": "string"
                  },
                  "artifactType": {
                    "type": "string",
                    "enum": [
                      "requirements",
                      "stories",
                      "activityDiagrams",
                      "useCaseDiagrams",
                      "sequenceDiagrams",
                      "classDiagrams",
                      "designPatterns",
                      "mockups"
                    ]
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment saved successfully"
          },
          "400": {
            "description": "Invalid artifact type"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to comment on this project"
          }
        }
      }
    },
    "/api/reviews/general-comment/{projectId}/{artifactType}": {
      "get": {
        "summary": "Get general comment for artifact type",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project ID"
          },
          {
            "in": "path",
            "name": "artifactType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "requirements",
                "stories",
                "activityDiagrams",
                "useCaseDiagrams",
                "sequenceDiagrams",
                "classDiagrams",
                "designPatterns",
                "mockups"
              ]
            },
            "description": "Type of artifact category"
          }
        ],
        "responses": {
          "200": {
            "description": "General comment for the artifact type"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to view comments for this project"
          }
        }
      }
    },
...
  },
...
}